+++ src/sidebar/ProjectComponent.ts
@@ -3,6 +3,7 @@ import { TreeDataProvider, TreeItem } from 'vscode'
 import { selectModule, selectModuleGlobal } from '../analysis/ModuleSelection'
 import { Symbol } from '../analysis/Symbol'
 import { ext } from '../extension'
+import { ConfigObject } from '../lib/libconfig'
 import {
   CommandNode,
   EditorButton,
@@ -11,6 +12,7 @@ import {
   ViewComponent,
 } from '../lib/libconfig'
 import { InstancesView, InstanceViewItem, ModuleItem } from './InstancesView'
+import path from 'path'
 
 export abstract class HierItem {
   getPath(): string {
@@ -103,6 +105,16 @@ export class InstanceItem extends ScopeItem {
     let item = await super.getTreeItem()
     item.contextValue = 'Module'
     item.iconPath = new vscode.ThemeIcon('chip')
+    if (this.definition && ext.project.showFileName.getValue()) {
+      item.description = '(' + path.basename(this.definition.doc.fileName) + ')'
+    }
+    if (this.definition && this.hasChildren() && ext.project.showChildCount.getValue()) {
+      let instNum = this.definition.children.filter(
+        (child) => child.type === 'instance').length
+      if (instNum > 0){
+        item.description += ' (' + instNum + ')'
+      }
+    }
     return item
   }
 
@@ -192,6 +204,19 @@ export class ProjectComponent
   // Hierarchy Tree
   private _onDidChangeTreeData: vscode.EventEmitter<void> = new vscode.EventEmitter<void>()
   readonly onDidChangeTreeData: vscode.Event<void> = this._onDidChangeTreeData.event
+
+  showFileName: ConfigObject<boolean> = new ConfigObject({
+    default: false,
+    description:
+      'Show the file name in which the module is declared in hierarchy view',
+  })
+
+  showChildCount: ConfigObject<boolean> = new ConfigObject({
+    default: false,
+    description:
+    'Show the count of child elements for each module in hierarchy view',
+  })
+
   treeView: vscode.TreeView<HierItem> | undefined
 
   // Instances Index